Moved Staking integration tests to the integration tests project.
