+++ src/Stratis.Bitcoin.IntegrationTests/NodeBuilder.cs
@@ -200,6 +200,28 @@ namespace Stratis.Bitcoin.IntegrationTests
             return builder.BuildAndRegister();
         }
 
+        /// <summary>
+        /// Builds a node with POS miner and RPC enabled.
+        /// </summary>
+        /// <param name="dir">Data directory that the node should use.</param>
+        /// <returns>Interface to the newly built node.</returns>
+        /// <remarks>Currently the node built here does not actually stake as it has no coins in the wallet,
+        /// but all the features required for it are enabled.</remarks>
+        public static IFullNode BuildStakingNode(string dir, bool staking = true)
+        {
+            NodeSettings nodeSettings = NodeSettings.FromArguments(new string[] { $"-datadir={dir}", $"-stake={(staking ? 1 : 0)}", "-walletname=dummy", "-walletpassword=dummy" });
+            var fullNodeBuilder = new FullNodeBuilder(nodeSettings);
+            IFullNode fullNode = fullNodeBuilder
+                .UseStratisConsensus()
+                .UseBlockStore()
+                .UseMempool()
+                .UseWallet()
+                .AddPowPosMining()
+                .AddRPC()
+                .Build();
+
+            return fullNode;
+        }
     }
 
     public class StratisBitcoinPowRunner : INodeRunner
